@@ -1,14 +1,18 @@
 module github.com/serious-snow/govm
 
-go 1.18
+go 1.22
+
+toolchain go1.22.6
 
 require (
-	github.com/cheggaaa/pb/v3 v3.1.4
-	github.com/fatih/color v1.15.0
-	github.com/go-ole/go-ole v1.2.6
+	github.com/briandowns/spinner v1.23.1
+	github.com/cheggaaa/pb/v3 v3.1.5
+	github.com/fatih/color v1.17.0
+	github.com/go-ole/go-ole v1.3.0
+	github.com/google/go-github/v66 v66.0.0
 	github.com/manifoldco/promptui v0.9.0
 	github.com/urfave/cli/v3 v3.0.0-alpha4
-	golang.org/x/sys v0.10.0
+	golang.org/x/sys v0.26.0
 	gopkg.in/yaml.v3 v3.0.1
 )
 
@@ -16,10 +20,12 @@ require (
 	github.com/VividCortex/ewma v1.2.0 // indirect
 	github.com/chzyer/readline v0.0.0-20180603132655-2972be24d48e // indirect
 	github.com/cpuguy83/go-md2man/v2 v2.0.2 // indirect
+	github.com/google/go-querystring v1.1.0 // indirect
 	github.com/mattn/go-colorable v0.1.13 // indirect
-	github.com/mattn/go-isatty v0.0.19 // indirect
+	github.com/mattn/go-isatty v0.0.20 // indirect
 	github.com/mattn/go-runewidth v0.0.15 // indirect
 	github.com/rivo/uniseg v0.4.4 // indirect
 	github.com/russross/blackfriday/v2 v2.1.0 // indirect
 	github.com/xrash/smetrics v0.0.0-20201216005158-039620a65673 // indirect
+	golang.org/x/term v0.1.0 // indirect
 )
